Refactor implementation of SoftwareBF16
