+++ mlir/test/mlir-miopen-driver/e2e_for_pr/bf16.mlir
@@ -1,0 +1,4 @@
+// RUN: miopen-gen -p -t f16 %pv %xdlops | mlir-miopen-driver -c | mlir-rocm-runner --shared-libs=%rocm_wrapper_library_dir/librocm-runtime-wrappers%shlibext,%linalg_test_lib_dir/libmlir_runner_utils%shlibext --entry-point-result=void | FileCheck %s --check-prefix=CHECK_BF16
+
+// CHECK_BF16: Unranked Memref base@ = 0x{{.*}} rank = 1 offset = 0 sizes = [1] strides = [1] data =
+// CHECK_BF16: [1]
